+++ client/src/app/models/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+    email: string;
+    token: string;
+}
