+++ client/src/feature/account/accountSlice.ts
@@ -1,0 +1,52 @@
+import { createAsyncThunk, createSlice, isAnyOf } from "@reduxjs/toolkit";
+import { FieldValues } from "react-hook-form";
+import agent from "../../app/api/agent";
+import { User } from "../../app/models/user";
+
+interface AccountState {
+    user: User | null;
+}
+
+const initialState: AccountState = {
+    user: null
+}
+
+export const SignInUser = createAsyncThunk<User, FieldValues>(
+    'account/signInUser',
+    async (data, thunkAPI) => {
+        try {
+            const user = await agent.Account.login(data);
+            localStorage.setItem('user', JSON.stringify(user));
+            return user;
+        } catch (error) {
+            return thunkAPI.rejectWithValue({error: error})
+        }
+    }
+)
+
+export const fetchCurrentUser = createAsyncThunk<User>(
+    'account/signInUser',
+    async (_, thunkAPI) => {
+        try {
+            const user = await agent.Account.currentUser();
+            localStorage.setItem('user', JSON.stringify(user));
+            return user;
+        } catch (error) {
+            return thunkAPI.rejectWithValue({error: error})
+        }
+    }
+)
+
+export const accountSlice = createSlice({
+    name: 'account',
+    initialState,
+    reducers: {},
+    extraReducers: (builder => {
+        builder.addMatcher(isAnyOf(SignInUser.fulfilled, fetchCurrentUser.fulfilled), (state, action) => {
+            state.user = action.payload;
+        });
+        builder.addMatcher(isAnyOf(SignInUser.rejected, fetchCurrentUser.rejected), (state, action) => {
+            console.log(action.payload);
+        });
+    })
+})
